fast-api/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
easyocr
python-multipart